requirements.txt
opencv-python
numpy
matplotlib
Pillow
tqdm